pyproject.toml
[tool.poetry]
name = "qmunity"
version = "0.1.0"
description = ""
authors = ["bitizen248 <van4011@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.88.0"
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.2"}
uvicorn = {extras = ["standard"], version = "^0.20.0"}
aerich = "^0.7.1"
bcrypt = "^4.0.1"
motor = "^3.1.1"



[tool.poetry.group.dev.dependencies]
black = "^22.10.0"

[tool.aerich]
tortoise_orm = "qmunity.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
